ThalAI-App-backup/ml-service/requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
numpy
joblib
python-dateutil
